Save of Lic Rule set

+++ committee/src/main/java/in/cdac/university/committee/bean/LicCommitteeRuleSetDtlBean.java
@@ -14,11 +14,11 @@ import lombok.ToString;
 public class LicCommitteeRuleSetDtlBean {
 
 	
-	private long unumComRsDtlId;
+	private Long unumComRsDtlId;
 	
 	private Integer unumComrolsno;
 	
-	private Integer unumComRsId;
+	private Long unumComRsId;
 	
 	private Integer unumCtypeid;
