+++ committee/src/main/java/in/cdac/university/committee/entity/GbltLicCommitteeRuleSetMstPk.java
@@ -7,7 +7,7 @@ import javax.persistence.Column;
 public class GbltLicCommitteeRuleSetMstPk implements Serializable {
 
 	@Column(name="unum_com_rs_id")
-	private Integer unumComRsId;
+	private Long unumComRsId;
 
 	@Column(name="unum_isvalid")
 	private Integer unumIsValid;
